feat: abstract CubeView to Add & Search Handler (#498)
* feat: abstract CubeView to Add Handler

* feat: add readable and writable memcube-ids

* feat: multi-cube search router
